Updated asizeof.py for Python 3.13 and __slots__ issue
The size of instances and dicts in Python 3.13+ respectively 3.6+ is much smaller, up to 5 times or more. This asizeof version 25.07.27 has been updated to correctly size those objects. See issue #170 for more details.

Yet another fix for issue #176 with sizing __slots__ attributes.

Other changes are updates to the basic sizes, a check for running with/out the GIL enabled and mostly cosmetic ones.

NOTE, the test_asizeof.py module must be updated as well to 25.07.27.